Remove potentially confusing EOF line from snippet
Our API getting started guide includes a go snippet that ends with
"EOF," which is not a Go keyword. If the reader isn't familiar with
Go but wants to follow this guide, the Go compiler will return a syntax
error.

This change puts "EOF" in a comment so this won't happen.